app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
bcrypt
python-dotenv
sqlalchemy
psycopg2-binary
opencv-python
pandas